requirements.txt
git+https://github.com/haven-jeon/PyKoSpacing.git
konlpy
python-Levenshtein
